requirements.txt
# Data & ML
pandas
numpy
scikit-learn
lightgbm
shap
dice-ml
imbalanced-learn
category_encoders

# Backend
fastapi
uvicorn
pydantic
python-multipart
python-dotenv  

# Frontend
streamlit
plotly
requests
httpx